Implement EditDialog component with form inputs
Created a new component EditDialog that renders a dialog box with
form inputs to allow users to edit the information of a post.
The component uses useState hook to manage the state of the edited post
information. It also includes event handlers for input changes and
buttons to confirm or cancel editing.