Create NavBar component with conditional rendering
This commit adds a NavBar component to the project, which displays
a navigation bar with links to the home page, sign up and log in pages,
and the option to add a new post. The component uses conditional
rendering to display different links depending on whether the user
is logged in or not. If the user is logged in, the component displays
a link to the add post page and a welcome message with the user's name.
If the user is not logged in, the component displays links to
the sign up and log in pages.